primjer_projekta(opis)/kod/python/requirements.txt
pyserial==3.5
PySimpleGUI==4.34.0
